Updated main.

+++ lib/include/syscalls.h
@@ -19,7 +19,7 @@ unsigned char *read_frame_buffer(size_t buffer_size);
 
 unsigned char *read_frame_buffer_with_offset(size_t buffer_size, off_t offset, off_t pa_offset);
 
-int write_to_frame_buffer(unsigned short *write_buffer, size_t num_bytes);
+int write_to_frame_buffer(unsigned short *write_buffer, int num_bytes);
 
 void exit_graphics();// ioctl
